LayoutTests/inspector/unit-tests/resource-collection-expected.txt

Testing all methods of ResourceCollection.


== Running test suite: ResourceCollection
-- Running test case: WebInspector.ResourceCollection.verifierForType
["two"]

-- Running test case: WebInspector.ResourceCollection.prototype.resourceCollectionForType
["one","three"]
["one","three"]
PASS: resourceCollectionForType will return `this` if a type is specified.

-- Running test case: WebInspector.ResourceCollection.prototype._resourceTypeDidChange
["one"]
[]
["one"]
["one"]
["one"]
[]
